correctly deserialise the 'editMode' attribute
The Happeo Custom Widget SDK sets the attribute as `editmode="false"` which, for a boolean property, by default gets deserialized to `true`. This is the standard behavior for boolean properties on HTML elements, e.g. `disabled` in form inputs behaves in the same way. In order to accommodate the non-standard attribute/property conversion used by the Happeo Custom Widget SDK, this example uses a custom attribute -> property converter.